requirements.txt
openai==0.27.8
docx2txt==0.8
langchain==0.0.225
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.24.0
pdfplumber
